+++ tasks/mincer.js
@@ -13,6 +13,7 @@ module.exports = function(grunt) {
   'use strict';
 
   var mince = require('./lib/mince').init(grunt).mince;
+  var _ = grunt.util._;
 
   grunt.registerMultiTask('mince', 'Use mincer to concatenate your files.', function () {
     function toArray(strOrArray) {
@@ -23,13 +24,18 @@ module.exports = function(grunt) {
       include = toArray(options.include),
       src = options.src || this.target + '.js',
       dest = options.dest || path.join(options.destDir, this.target + '.js'),
+      // default `writeAsset` to true since that is the original behaviour of grunt-mincer
+      writeAsset = (options.writeAsset === void 0) || options.writeAsset,
+      onCompile = options.onCompile || null,
       done = this.async();
 
-    grunt.log.write('Generating file ' + dest.cyan + '...');
-    mince(src, dest, include, options.configure, function(err) {
+    mince(src, dest, include, options.configure, writeAsset, function(err, asset) {
       if (err) {
         grunt.warn(err);
       } else {
+        if (grunt.util.kindOf(onCompile) === 'function') {
+          onCompile(_.merge({ asset: asset }, options));
+        }
         grunt.log.ok();
       }
       done();